cloud_sql_python_code/cloud_sql_python_in_cloud_shell/requirements.txt
PyMySQL==1.0.2
python-dotenv==0.20.0
cryptography==37.0.4